chore: Update Prettier and ESLint configurations
This commit updates the Prettier configuration to include the "endOfLine" option set to "lf" and the ESLint configuration to include the rule "@typescript-eslint/no-unused-vars" set to "warn" and disable the "prettier/prettier" rule.

It also adds new dependencies to the package.json file: "@nestjs/config" version "^3.2.3" and "@nestjs/mongoose" version "^10.0.10". Additionally, it updates the "mongoose" dependency to version "^8.6.0".

In the app.module.ts file, it imports the "BookModule" and adds the "ConfigModule" with the options "envFilePath: '.env'" and "isGlobal: true". It also adds the "MongooseModule" with the "DB_url" environment variable as the argument to the "forRoot" method.

The "src/app.controller.spec.ts" file has been deleted.

```